mito-ai: add styling
