+++ mito-ai/style/NotebookFooter.css
@@ -1,0 +1,143 @@
+/* Main container */
+.notebook-footer-container {
+    position: relative;
+    width: 100%;
+    padding: 12px 16px;
+    background-color: var(--jp-layout-color1);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 8px;
+}
+
+/* Button row */
+.button-row {
+    display: flex;
+    gap: 12px;
+    align-items: center;
+    justify-content: center;
+}
+
+/* Footer buttons */
+.footer-button {
+    min-width: 80px;
+    padding: 8px 12px;
+    border: 1px solid var(--jp-border-color2);
+    border-radius: 4px;
+    background-color: var(--jp-layout-color2);
+    color: var(--jp-ui-font-color1);
+    font-size: 12px;
+    font-family: var(--jp-ui-font-family);
+    cursor: pointer;
+    transition: all 0.2s ease;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 4px;
+}
+
+.footer-button:hover {
+    background-color: var(--jp-layout-color3);
+    border-color: var(--jp-brand-color1);
+}
+
+.footer-button:active {
+    background-color: var(--jp-brand-color1);
+    color: white;
+}
+
+/* Button content */
+.button-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 4px;
+}
+
+/* Button icon */
+.button-icon {
+    font-size: 16px;
+    line-height: 1;
+}
+
+/* Button label */
+.button-label {
+    font-size: 11px;
+    font-weight: 500;
+    line-height: 1;
+}
+
+/* Python button specific styling */
+.python-button .button-icon {
+    color: #3776ab;
+}
+
+/* Text button specific styling */
+.text-button .button-icon {
+    color: var(--jp-ui-font-color2);
+}
+
+/* Cell info */
+.cell-info {
+    font-size: 11px;
+    color: var(--jp-ui-font-color2);
+    font-family: var(--jp-ui-font-family);
+    text-align: center;
+}
+
+/* Legacy styles for backward compatibility */
+.add-button {
+    flex: 1;
+    padding: 12px;
+    border: none;
+    border-radius: 8px;
+    background-color: #4CAF50;
+    color: white;
+    font-size: 14px;
+    font-weight: bold;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+}
+
+.add-button:hover {
+    background-color: #45a049;
+}
+
+.dropdown-toggle {
+    padding: 12px 8px;
+    border: none;
+    border-radius: 8px;
+    background-color: #4CAF50;
+    color: white;
+    cursor: pointer;
+    font-size: 12px;
+}
+
+.dropdown-menu {
+    position: absolute;
+    bottom: 100%;
+    right: 0;
+    background-color: white;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.1);
+    z-index: 1000;
+    min-width: 150px;
+}
+
+.dropdown-item {
+    width: 100%;
+    padding: 8px 12px;
+    border: none;
+    background: none;
+    text-align: left;
+    cursor: pointer;
+}
+
+.dropdown-item:hover {
+    background-color: #f5f5f5;
+}
